requirements.txt
blinker==1.9.0
click==8.1.7
contourpy==1.3.1
cycler==0.12.1
Flask==3.1.0
fonttools==4.55.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.2
numpy==2.1.3
packaging==24.2
pillow==11.0.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
scikit-learn==1.5.2
scipy==1.14.1
six==1.16.0
threadpoolctl==3.5.0
Werkzeug==3.1.3
